update logging on provider webrtc signalling server
