+++ hub/gads-ui/src/components/DeviceControl/DeviceControl.js
@@ -47,8 +47,8 @@ export default function DeviceControl() {
             wsType = 'wss'
         }
 
-        let socketUrl = `${wsType}://${window.location.host}/devices/control/${udid}/in-use`
-        // let socketUrl = `${wsType}://192.168.1.41:10000/devices/control/${udid}/in-use`
+        // let socketUrl = `${wsType}://${window.location.host}/devices/control/${udid}/in-use`
+        let socketUrl = `${wsType}://192.168.1.41:10000/devices/control/${udid}/in-use`
 
         in_use_socket = new WebSocket(socketUrl)
         in_use_socket.onopen = () => {
